Bring execution code under a single function
