+++ lib/tasks/install-blueprint.js
@@ -10,7 +10,7 @@ const merge = require('ember-cli-lodash-subset').merge;
 const SilentError = require('silent-error');
 const npa = require('npm-package-arg');
 const lintFix = require('../utilities/lint-fix');
-const { commands } = require('../utilities/package-managers');
+const { NPM, runCommand } = require('../utilities/package-managers');
 
 const logger = require('heimdalljs-logger')('ember-cli:tasks:install-blueprint');
 
@@ -171,7 +171,7 @@ class InstallBlueprintTask extends Task {
 
     this._copyNpmrc(cwd);
 
-    return commands.npm.invoke('install', '--no-package-lock', '--save', name, { cwd });
+    return runCommand(NPM, 'install', ['--no-package-lock', '--save', name], { cwd });
   }
 
   _handleNpmInstallModuleError(error) {
