+++ lib/tasks/npm-task.js
@@ -7,7 +7,7 @@ const findUp = require('find-up');
 const semver = require('semver');
 const SilentError = require('silent-error');
 const isYarnProject = require('../utilities/is-yarn-project');
-const { commands } = require('../utilities/package-managers');
+const { NPM, YARN, PNPM, runCommand } = require('../utilities/package-managers');
 
 const logger = require('heimdalljs-logger')('ember-cli:npm-task');
 
@@ -33,21 +33,6 @@ class NpmTask extends Task {
     return this.packageManager.name;
   }
 
-  npm(args) {
-    logger.info('npm: %j', args);
-    return commands.npm.invoke(args, { preferLocal: false });
-  }
-
-  yarn(args) {
-    logger.info('yarn: %j', args);
-    return commands.yarn.invoke(args, { preferLocal: false });
-  }
-
-  pnpm(args) {
-    logger.info('pnpm: %j', args);
-    return commands.pnpm.invoke(args, { preferLocal: false });
-  }
-
   hasYarnLock() {
     return isYarnProject(this.project.root);
   }
@@ -58,7 +43,7 @@ class NpmTask extends Task {
 
   async checkYarn() {
     try {
-      let result = await this.yarn(['--version']);
+      let result = await runCommand(YARN, ['--version']);
       let version = result.stdout;
 
       if (semver.gte(version, '2.0.0')) {
@@ -85,7 +70,7 @@ class NpmTask extends Task {
 
   async checkPNPM() {
     try {
-      let result = await this.pnpm(['--version']);
+      let result = await runCommand(PNPM, ['--version']);
       let version = result.stdout;
 
       logger.info('pnpm --version: %s', version);
@@ -107,7 +92,7 @@ class NpmTask extends Task {
 
   async checkNpmVersion() {
     try {
-      let result = await this.npm(['--version']);
+      let result = await runCommand(NPM, ['--version']);
       let version = result.stdout;
       logger.info('npm --version: %s', version);
 
@@ -159,17 +144,17 @@ class NpmTask extends Task {
    * @return {Promise}
    */
   async findPackageManager(packageManager = null) {
-    if (packageManager === 'yarn') {
+    if (packageManager === YARN) {
       logger.info('yarn requested -> trying yarn');
       return this.checkYarn();
     }
 
-    if (packageManager === 'npm') {
+    if (packageManager === NPM) {
       logger.info('npm requested -> using npm');
       return this.checkNpmVersion();
     }
 
-    if (packageManager === 'pnpm') {
+    if (packageManager === PNPM) {
       logger.info('pnpm requested -> using pnpm');
       return this.checkPNPM();
     }
@@ -218,143 +203,12 @@ class NpmTask extends Task {
     ui.startProgress(chalk.green(startMessage));
 
     try {
-      if (this.packageManager.name === 'yarn') {
-        let args = this.toYarnArgs(this.command, options);
-        await this.yarn(args);
-      } else if (this.packageManager.name === 'pnpm') {
-        let args = this.toPNPMArgs(this.command, options);
-        await this.pnpm(args);
-      } else {
-        let args = this.toNpmArgs(this.command, options);
-        await this.npm(args);
-
-        // as of 2018-10-09 npm 5 and 6 _break_ the hierarchy of `node_modules`
-        // after a `npm install foo` (deletes files/folders other than
-        // what was directly installed) in some circumstances, see:
-        //
-        // * https://github.com/npm/npm/issues/16853
-        // * https://github.com/npm/npm/issues/17379
-        //
-        // this ensures that we run a full `npm install` **after** any `npm
-        // install foo` runs to ensure that we have a fully functional
-        // node_modules hierarchy
-        let npmVersion = this.packageManager.version;
-        if (npmVersion && semver.lt(npmVersion, '5.7.1')) {
-          await this.npm(['install']);
-        }
-      }
+      await runCommand(this.packageManager.name, this.command, options);
     } finally {
       ui.stopProgress();
     }
 
     ui.writeLine(chalk.green(completeMessage));
-  }
-
-  toNpmArgs(command, options) {
-    let args = [command];
-
-    if (options.save) {
-      args.push('--save');
-    }
-
-    if (options['save-dev']) {
-      args.push('--save-dev');
-    }
-
-    if (options['save-exact']) {
-      args.push('--save-exact');
-    }
-
-    if ('optional' in options && !options.optional) {
-      args.push('--no-optional');
-    }
-
-    if (options.verbose) {
-      args.push('--loglevel', 'verbose');
-    } else {
-      args.push('--loglevel', 'error');
-    }
-
-    if (options.packages) {
-      args = args.concat(options.packages);
-    }
-
-    return args;
-  }
-
-  toYarnArgs(command, options) {
-    let args = [];
-
-    if (command === 'install') {
-      if (options.save) {
-        args.push('add');
-      } else if (options['save-dev']) {
-        args.push('add', '--dev');
-      } else if (options.packages) {
-        throw new Error(`npm command "${command} ${options.packages.join(' ')}" can not be translated to Yarn command`);
-      } else {
-        args.push('install');
-      }
-
-      if (options['save-exact']) {
-        args.push('--exact');
-      }
-
-      if ('optional' in options && !options.optional) {
-        args.push('--ignore-optional');
-      }
-    } else if (command === 'uninstall') {
-      args.push('remove');
-    } else {
-      throw new Error(`npm command "${command}" can not be translated to Yarn command`);
-    }
-
-    if (options.verbose) {
-      args.push('--verbose');
-    }
-
-    if (options.packages) {
-      args = args.concat(options.packages);
-    }
-
-    // Yarn v2 defaults to non-interactive
-    // with an optional -i flag
-
-    if (semver.lt(this.packageManager.version, '2.0.0')) {
-      args.push('--non-interactive');
-    }
-
-    return args;
-  }
-
-  toPNPMArgs(command, options) {
-    let args = [];
-
-    if (command === 'install') {
-      if (options.save) {
-        args.push('add');
-      } else if (options['save-dev']) {
-        args.push('add', '--save-dev');
-      } else if (options.packages) {
-        throw new Error(`npm command "${command} ${options.packages.join(' ')}" can not be translated to pnpm command`);
-      } else {
-        args.push('install');
-      }
-
-      if (options['save-exact']) {
-        args.push('--save-exact');
-      }
-    } else if (command === 'uninstall') {
-      args.push('remove');
-    } else {
-      throw new Error(`npm command "${command}" can not be translated to pnpm command`);
-    }
-
-    if (options.packages) {
-      args = args.concat(options.packages);
-    }
-
-    return args;
   }
 
   formatStartMessage(/* packages */) {
